Sites

+++ src/resolvers/ticket/ticket.resolver.ts
@@ -5,7 +5,6 @@ import {
 } from '@nestjs/common';
 import { Args, Int, Mutation, Query, Resolver } from '@nestjs/graphql';
 import { GqlAuthGuard } from '../../guards/gql-auth.guard';
-import { Roles } from 'src/decorators/roles.decorator';
 import { RolesGuard } from 'src/guards/roles.guard';
 import { TicketService } from 'src/services/ticket.service';
 import { Ticket } from 'src/models/ticket.model';
@@ -23,6 +22,8 @@ import * as moment from 'moment';
 import { TicketStatusCountHistory } from 'src/models/ticket-status-count-history';
 import { RedisCacheService } from 'src/redisCache.service';
 import { AgentQueue } from 'src/models/agent-queue.model';
+import { SiteService } from 'src/services/site.service';
+import { RoleEnum } from 'src/common/enums/roles';
 
 @Resolver(() => Ticket)
 @UseGuards(GqlAuthGuard, RolesGuard)
@@ -31,36 +32,43 @@ export class TicketResolver {
     private ticketService: TicketService,
     private prisma: PrismaService,
     private userService: UserService,
-    private readonly redisCacheService: RedisCacheService
+    private readonly redisCacheService: RedisCacheService,
+    private siteService: SiteService
   ) {}
 
   @Mutation(() => Int)
   async createTicket(
     @UserEntity() user: User,
     @Args('title') title: string,
-    @Args('body') body: string
+    @Args('body') body: string,
+    @Args('siteId') siteId: number
   ): Promise<number> {
-    return await this.ticketService.createTicket(user, title, body);
+    await this.siteService.checkSiteAccess(user.id, siteId);
+    return await this.ticketService.createTicket(user, title, body, siteId);
   }
 
-  @Roles('Admin', 'Agent')
   @Mutation(() => String)
   async setTicketPriority(
     @UserEntity() user: User,
     @Args('ticketId') ticketId: number,
     @Args('priority', { type: () => Priority }) priority: Priority
   ): Promise<string> {
+    const ticket = await this.prisma.ticket.findFirst({
+      where: { id: ticketId },
+    });
+    await this.userService.checkAdminOrAgent(user.id, ticket.id);
     await this.ticketService.setTicketPriority(user, ticketId, priority);
     return `Ticket priority set to ${priority}.`;
   }
 
-  @Roles('Admin', 'Agent')
   @Mutation(() => String)
   async setTicketStatus(
     @UserEntity() user: User,
     @Args('ticketId') id: number,
     @Args('status', { type: () => Status }) status: Status
   ): Promise<string> {
+    const ticket = await this.prisma.ticket.findFirst({ where: { id: id } });
+    await this.userService.checkAdminOrAgent(user.id, ticket.id);
     await this.ticketService.setTicketStatus(user, id, status);
     return `Ticket status set to ${status}.`;
   }
@@ -101,80 +109,119 @@ export class TicketResolver {
     return `Successfully removed follower from ticket.`;
   }
 
-  @Roles('Admin', 'Agent')
   @Mutation(() => String)
   async assignAgents(
     @UserEntity() user: User,
     @Args('ticketId') ticketId: number,
     @Args('agentIds', { type: () => [Int] }) agentIds: number[]
   ): Promise<string> {
-    await this.ticketService.assignAgents(user, ticketId, agentIds);
+    const ticket = await this.prisma.ticket.findFirst({
+      where: { id: ticketId },
+    });
+    const isAdmin = await this.userService.isAdmin(user.id, ticket.siteId);
+    const isAgent = await this.userService.isAgent(user.id, ticket.siteId);
+    if (!isAdmin && !isAgent) {
+      throw new UnauthorizedException('You do not have access to this ticket.');
+    }
+    await this.ticketService.assignAgents(
+      user,
+      ticketId,
+      agentIds,
+      isAdmin,
+      ticket.siteId
+    );
     return `Successfully assigned agent${
       agentIds.length > 1 ? 's' : ''
     } to ticket.`;
   }
 
-  @Roles('Admin', 'Agent')
   @Mutation(() => String)
   async setOwner(
     @UserEntity() user: User,
     @Args('ticketId') ticketId: number,
     @Args('agentId') agentId: number
   ): Promise<string> {
-    await this.ticketService.setOwner(user, ticketId, agentId);
+    const ticket = await this.prisma.ticket.findFirst({
+      where: { id: ticketId },
+    });
+    await this.userService.checkAdminOrAgent(user.id, ticket.id);
+    await this.ticketService.setOwner(user, ticketId, agentId, ticket.siteId);
     return `Successfully set new owner of ticket.`;
   }
 
-  @Roles('Admin', 'Agent')
   @Mutation(() => String)
   async unassignAgent(
     @UserEntity() user: User,
     @Args('ticketId') ticketId: number,
     @Args('agentId') agentId: number
   ): Promise<string> {
-    await this.ticketService.unassignAgent(user, ticketId, agentId);
+    const ticket = await this.prisma.ticket.findFirst({
+      where: { id: ticketId },
+    });
+    await this.userService.checkAdminOrAgent(user.id, ticket.id);
+    await this.ticketService.unassignAgent(
+      user,
+      ticketId,
+      agentId,
+      ticket.siteId
+    );
     return `Successfully unassigned agent to ticket.`;
   }
 
-  @Roles('Admin', 'Agent')
   @Mutation(() => String)
   async addChecklistItem(
     @UserEntity() user: User,
     @Args('ticketId') ticketId: number,
     @Args('description') description: string
   ): Promise<string> {
+    const ticket = await this.prisma.ticket.findFirst({
+      where: { id: ticketId },
+    });
+    await this.userService.checkAdminOrAgent(user.id, ticket.id);
     await this.ticketService.createChecklistItem(user, ticketId, description);
     return `Added checklist item to ticket.`;
   }
 
-  @Roles('Admin', 'Agent')
   @Mutation(() => String)
   async editChecklistItem(
     @UserEntity() user: User,
     @Args('id') id: number,
     @Args('description') description: string
   ): Promise<string> {
+    const checklistItem = await this.prisma.checklistItem.findFirst({
+      where: { id },
+      include: { ticket: true },
+    });
+    await this.userService.checkAdminOrAgent(user.id, checklistItem.ticket.id);
     await this.ticketService.editChecklistItem(user, id, description);
     return `Checklist item updated.`;
   }
 
-  @Roles('Admin', 'Agent')
   @Mutation(() => String)
   async toggleChecklistItem(
     @UserEntity() user: User,
     @Args('id') id: number,
     @Args('complete') complete: boolean
   ): Promise<string> {
+    const checklistItem = await this.prisma.checklistItem.findFirst({
+      where: { id },
+      include: { ticket: true },
+    });
+    await this.userService.checkAdminOrAgent(user.id, checklistItem.ticket.id);
     await this.ticketService.toggleChecklistItem(user, id, complete);
     return `Checklist item updated.`;
   }
 
-  @Roles('Admin', 'Agent')
   @Mutation(() => String)
   async deleteChecklistItem(
     @UserEntity() user: User,
     @Args('id') id: number
   ): Promise<string> {
+    const checklistItem = await this.prisma.checklistItem.findFirst({
+      where: { id },
+      include: { ticket: true },
+    });
+    await this.userService.checkAdminOrAgent(user.id, checklistItem.ticket.id);
     await this.ticketService.deleteChecklistItem(user, id);
     return `Checklist item deleted.`;
   }
@@ -195,18 +242,16 @@ export class TicketResolver {
     @UserEntity() user: User,
     @Args() args: TicketConnectionArgs
   ): Promise<PaginatedTickets> {
-    args.self = true;
     args.createdById = user.id;
     return await this.ticketService.getTicketsWithPagination(user, args);
   }
 
-  @Roles('Admin', 'Agent')
   @Query(() => PaginatedTickets)
   async tickets(
     @UserEntity() user: User,
     @Args() args: TicketConnectionArgs
   ): Promise<PaginatedTickets> {
-    args.self = false;
+    args.all = true;
     if (args.createdByUserId) {
       const createdBy = await this.prisma.user.findFirst({
         where: { userId: args.createdByUserId },
@@ -220,13 +265,11 @@ export class TicketResolver {
     return await this.ticketService.getTicketsWithPagination(user, args);
   }
 
-  @Roles('Agent')
   @Query(() => PaginatedTickets)
   async assignedTickets(
     @UserEntity() user: User,
     @Args() args: TicketConnectionArgs
   ): Promise<PaginatedTickets> {
-    args.self = false;
     args.assignedToId = user.id;
     if (args.createdByUserId) {
       const createdBy = await this.prisma.user.findFirst({
@@ -246,7 +289,6 @@ export class TicketResolver {
     @UserEntity() user: User,
     @Args() args: TicketConnectionArgs
   ): Promise<PaginatedTickets> {
-    args.self = false;
     args.followingId = user.id;
     if (args.createdByUserId) {
       const createdBy = await this.prisma.user.findFirst({
@@ -277,7 +319,6 @@ export class TicketResolver {
     return await this.ticketService.ticket(user, ticketId);
   }
 
-  @Roles('Admin', 'Agent')
   @Mutation(() => String)
   async addTicketCategory(
     @UserEntity() user: User,
@@ -298,7 +339,6 @@ export class TicketResolver {
     }
   }
 
-  @Roles('Admin', 'Agent')
   @Mutation(() => String)
   async removeTicketCategory(
     @UserEntity() user: User,
@@ -321,19 +361,41 @@ export class TicketResolver {
 
   @Query(() => [TicketStatusCount])
   async ticketStatusCount(
-    @UserEntity() user: User
+    @UserEntity() user: User,
+    @Args('siteId', { nullable: true }) siteId?: number
   ): Promise<TicketStatusCount[]> {
-    const isAdminOrAgent = await this.userService.isAdminOrAgent(user.id);
-    let statusCounts = await this.redisCacheService.get(
-      `statusCount${isAdminOrAgent ? '' : `Self-${user.id}`}`
+    let type = 'Self';
+    let siteIds = [];
+    const sitesWithRoles = await this.siteService.sitesWithRoles(
+      user.id,
+      [RoleEnum.Admin, RoleEnum.Agent],
+      user
     );
+    if (sitesWithRoles.length > 0) {
+      if (siteId && sitesWithRoles.includes(siteId)) {
+        type = 'Site';
+        siteIds = [siteId];
+      } else {
+        type = 'OwnSites';
+        siteIds = sitesWithRoles;
+      }
+    }
+    const key = `statusCount${
+      type === 'Self'
+        ? `Self-${user.id}`
+        : type === 'Site'
+        ? `Site-${siteId}`
+        : `OwnSites-${user.id}`
+    }`;
+    let statusCounts = await this.redisCacheService.get(key);
     if (!statusCounts) {
       statusCounts = [];
       for (const status of Object.keys(Status) as Array<keyof typeof Status>) {
         const count = await this.prisma.ticket.count({
           where: {
             status: status,
-            createdById: isAdminOrAgent ? undefined : user.id,
+            createdById: type === 'Self' ? user.id : undefined,
+            siteId: type === 'Self' ? undefined : { in: siteIds },
           },
         });
         statusCounts.push({
@@ -341,21 +403,32 @@ export class TicketResolver {
           count,
         });
       }
-      await this.redisCacheService.setFor10Sec(
-        `statusCount${isAdminOrAgent ? '' : `Self-${user.id}`}`,
-        statusCounts
-      );
+      await this.redisCacheService.setFor10Sec(key, statusCounts);
     }
     return statusCounts;
   }
 
-  @Roles('Admin', 'Agent')
   @Query(() => [TicketStatusCountHistory])
   async ticketStatusCountHistory(
+    @UserEntity() user: User,
     @Args() args: TicketStatusHistoryConnectionArgs
   ): Promise<TicketStatusCountHistory[]> {
     let { statuses } = args;
-    const { from, to } = args;
+    const { from, to, siteId } = args;
+    const sitesWithRoles = await this.siteService.sitesWithRoles(
+      user.id,
+      [RoleEnum.Admin, RoleEnum.Agent],
+      user
+    );
+    if (sitesWithRoles.length === 0) {
+      throw new UnauthorizedException(
+        'You do not have priviledged access to any sites.'
+      );
+    }
+    let siteIds = sitesWithRoles;
+    if (siteId && sitesWithRoles.includes(siteId)) {
+      siteIds = [siteId];
+    }
     const today = moment();
     const fromDate = moment(from).startOf('day');
     const toDate = moment(to).endOf('day');
@@ -367,17 +440,17 @@ export class TicketResolver {
         (s) => Status[s]
       );
     }
-    let statusHistoryByDate = await this.redisCacheService.get(
-      `statusCountHistory-${statuses.join(',')}-${fromDate.format(
-        'DD-MMMM-YYYY'
-      )}-${toDate.format('DD-MMMM-YYYY')}`
-    );
+    const key = `statusCountHistory-${statuses.join(',')}-${siteIds.join(
+      ','
+    )}-${fromDate.format('DD-MMMM-YYYY')}-${toDate.format('DD-MMMM-YYYY')}`;
+    let statusHistoryByDate = await this.redisCacheService.get(key);
     if (!statusHistoryByDate) {
       statusHistoryByDate = [];
       const statusHistory = await this.prisma.ticketStatusHistory.findMany({
         where: {
           status: { in: statuses },
           createdAt: { gte: from, lte: toDate.toDate() },
+          siteId: { in: siteIds },
         },
       });
       const days = toDate.diff(fromDate, 'days') + 1;
@@ -396,30 +469,43 @@ export class TicketResolver {
           statusCounts,
         });
       }
-      await this.redisCacheService.setForDay(
-        `statusCountHistory-${statuses.join(',')}-${fromDate.format(
-          'DD-MMMM-YYYY'
-        )}-${toDate.format('DD-MMMM-YYYY')}`,
-        statusHistoryByDate
-      );
+      await this.redisCacheService.setForDay(key, statusHistoryByDate);
     }
     return statusHistoryByDate;
   }
 
-  @Roles('Admin', 'Agent')
   @Query(() => [AgentQueue])
-  async agentQueue(): Promise<AgentQueue[]> {
-    let agentQueue = await this.redisCacheService.get(`agentQueue`);
+  async agentQueue(
+    @UserEntity() user: User,
+    @Args('siteId', { nullable: true }) siteId?: number
+  ): Promise<AgentQueue[]> {
+    const sitesWithRoles = await this.siteService.sitesWithRoles(
+      user.id,
+      [RoleEnum.Admin, RoleEnum.Agent],
+      user
+    );
+    if (sitesWithRoles.length === 0) {
+      throw new UnauthorizedException(
+        'You do not have priviledged access to any sites.'
+      );
+    }
+    let siteIds = sitesWithRoles;
+    if (siteId && sitesWithRoles.includes(siteId)) {
+      siteIds = [siteId];
+    }
+    const key = `agentQueue-${siteIds.join(',')}`;
+    let agentQueue = await this.redisCacheService.get(key);
     if (!agentQueue) {
       agentQueue = [];
       const activeTickets = await this.prisma.ticket.findMany({
         where: {
           status: { in: ['Pending', 'Open', 'Reopened'] },
+          siteId: { in: siteIds },
         },
-        include: { ticketAssignments: true },
+        include: { ticketAssignments: true, site: true },
       });
       const allAgents = await this.prisma.user.findMany({
-        where: { roles: { some: { role: 'Agent' } } },
+        where: { roles: { some: { role: 'Agent', siteId: { in: siteIds } } } },
       });
       allAgents.forEach((agent) => {
         const tickets = activeTickets.filter((ticket) => {
@@ -433,7 +519,7 @@ export class TicketResolver {
           tickets,
         });
       });
-      await this.redisCacheService.setFor10Sec(`agentQueue`, agentQueue);
+      await this.redisCacheService.setFor10Sec(key, agentQueue);
     }
     return agentQueue;
   }
